Updated project to use Androidx instead of older Legacy android UI classes.
Updated eh Order model class to take a DocumentSnapshot instead of a DocumentReference since the model class is static and doesn't change automatically with Firebase.

Also modified the CookActivity, OrderListViewModel, and OrderListAdapter classes to use a LiveData list of orders.

Fixed a bug that caused the LiveData to be null when it first loads.